backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-dotenv==1.0.1
python-multipart==0.0.9
anthropic==0.34.2
openai-whisper==20231117
pydub==0.25.1
